refactor: internalize implementation packages
