+++ conditional.go
@@ -5,11 +5,11 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/buildkite/conditional/ast"
-	"github.com/buildkite/conditional/evaluator"
-	"github.com/buildkite/conditional/lexer"
-	"github.com/buildkite/conditional/object"
-	"github.com/buildkite/conditional/parser"
+	"github.com/buildkite/conditional/internal/ast"
+	"github.com/buildkite/conditional/internal/evaluator"
+	"github.com/buildkite/conditional/internal/lexer"
+	"github.com/buildkite/conditional/internal/object"
+	"github.com/buildkite/conditional/internal/parser"
 )
 
 // Validate parses expression for the selected Buildkite context.
