+++ internal/evaluator/benchmark_test.go
@@ -3,9 +3,9 @@ package evaluator
 import (
 	"testing"
 
-	"github.com/buildkite/conditional/lexer"
-	"github.com/buildkite/conditional/object"
-	"github.com/buildkite/conditional/parser"
+	"github.com/buildkite/conditional/internal/lexer"
+	"github.com/buildkite/conditional/internal/object"
+	"github.com/buildkite/conditional/internal/parser"
 )
 
 // An example of evaluating a single token to get a baseline of evaluator overhead
